move container to component folder
