+++ src/client/src/components/RightSidebar/About/index.tsx
