+++ src/client/src/components/RightSidebar/ProjectDetails/ProjectDetails.test.tsx
