+++ src/client/src/components/RightSidebar/ProjectDetails/index.tsx
