+++ src/client/src/components/RightSidebar/ServicesList/AppServiceSelection/styles.module.css
