+++ src/client/src/components/RightSidebar/ServicesList/CosmosDBSelection/index.tsx
