+++ src/client/src/components/RightSidebar/ServicesList/CosmosDBSelection/messages.ts
