+++ src/client/src/components/RightSidebar/ServicesList/CosmosDBSelection/styles.module.css
