+++ src/client/src/components/RightSidebar/ServicesList/strings.ts
